shensibackend/app/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
atlastk==0.13.3
attrs==23.2.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iso8601==1.1.0
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.25.0
yarl==1.9.4
captcha
itsdangerous
redis
python-dotenv